+++ src/main/java/com/hackathon/sprout/domain/chat/dto/request/ChatMessageCreateRequest.java
@@ -1,4 +1,4 @@
-package com.hackathon.sprout.domain.chat.dto;
+package com.hackathon.sprout.domain.chat.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
